// web/src/api/models/TypesAnnouncement.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesAnnouncement
 */
export interface TypesAnnouncement {
    /**
     * 
     * @type {number}
     * @memberof TypesAnnouncement
     */
    created_at: number;
    /**
     * 
     * @type {boolean}
     * @memberof TypesAnnouncement
     */
    dismissible: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof TypesAnnouncement
     */
    enabled: boolean;
    /**
     * 
     * @type {number}
     * @memberof TypesAnnouncement
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof TypesAnnouncement
     */
    level: TypesAnnouncementLevelEnum;
    /**
     * 
     * @type {string}
     * @memberof TypesAnnouncement
     */
    message: string;
    /**
     * 
     * @type {number}
     * @memberof TypesAnnouncement
     */
    updated_at: number;
}


/**
 * @export
 */
export const TypesAnnouncementLevelEnum = {
    Info: 'info',
    Warning: 'warning',
    Danger: 'danger'
} as const;
export type TypesAnnouncementLevelEnum = typeof TypesAnnouncementLevelEnum[keyof typeof TypesAnnouncementLevelEnum];


/**
 * Check if a given object implements the TypesAnnouncement interface.
 */
export function instanceOfTypesAnnouncement(value: object): value is TypesAnnouncement {
    if (!('created_at' in value) || value['created_at'] === undefined) return false;
    if (!('dismissible' in value) || value['dismissible'] === undefined) return false;
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('level' in value) || value['level'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('updated_at' in value) || value['updated_at'] === undefined) return false;
    return true;
}

export function TypesAnnouncementFromJSON(json: any): TypesAnnouncement {
    return TypesAnnouncementFromJSONTyped(json, false);
}

export function TypesAnnouncementFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesAnnouncement {
    if (json == null) {
        return json;
    }
    return {
        
        'created_at': json['created_at'],
        'dismissible': json['dismissible'],
        'enabled': json['enabled'],
        'id': json['id'],
        'level': json['level'],
        'message': json['message'],
        'updated_at': json['updated_at'],
    };
}

export function TypesAnnouncementToJSON(json: any): TypesAnnouncement {
    return TypesAnnouncementToJSONTyped(json, false);
}

export function TypesAnnouncementToJSONTyped(value?: TypesAnnouncement | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['created_at'],
        'dismissible': value['dismissible'],
        'enabled': value['enabled'],
        'id': value['id'],
        'level': value['level'],
        'message': value['message'],
        'updated_at': value['updated_at'],
    };
}